+++ include/mockturtle/views/cut_view.hpp
@@ -104,7 +104,7 @@ public:
     }
   }
 
-  template<typename _Ntk = Ntk, typename = std::enable_if_t<!std::is_same_v<_Ntk::signal, _Ntk::node>>>
+  template<typename _Ntk = Ntk, typename = std::enable_if_t<!std::is_same_v<typename _Ntk::signal, typename _Ntk::node>>>
   explicit cut_view( Ntk const& ntk, std::vector<signal> const& leaves, node const& root )
       : immutable_view<Ntk>( ntk ), _root( root )
   {
@@ -227,7 +227,7 @@ public:
 template<class T>
 cut_view(T const&, std::vector<node<T>> const&, node<T> const&) -> cut_view<T>;
 
-template<class T, typename = std::enable_if_t<!std::is_same_v<T::signal, T::node>>>
+template<class T, typename = std::enable_if_t<!std::is_same_v<typename T::signal, typename T::node>>>
 cut_view(T const&, std::vector<signal<T>> const&, node<T> const&) -> cut_view<T>;
 
 } /* namespace mockturtle */