+++ lib/percy/percy/encoders.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "encoders/knuth_encoder.hpp"
-#include "encoders/knuth_fence_encoder.hpp"
-#include "encoders/knuth_fence2_encoder.hpp"
-#include "encoders/knuth_dag_encoder.hpp"
-#include "encoders/epfl_encoder.hpp"
-#include "encoders/berkeley_encoder.hpp"
+#include "encoders/ssv_encoder.hpp"
+#include "encoders/ssv_fence_encoder.hpp"
+#include "encoders/ssv_fence2_encoder.hpp"
+#include "encoders/ssv_dag_encoder.hpp"
+#include "encoders/msv_encoder.hpp"
+#include "encoders/ditt_encoder.hpp"
 #include "encoders/partial_dag_encoder.hpp"
 #include "encoders/maj_encoder.hpp"