+++ lib/percy/percy/encoders/ssv_fence2_encoder.hpp
@@ -5,7 +5,7 @@
 
 namespace percy
 {
-    class knuth_fence2_encoder : public fence_encoder
+    class ssv_fence2_encoder : public fence_encoder
     {
     private:
         int level_dist[32]; // How many steps are below a certain level
@@ -22,11 +22,14 @@ namespace percy
 
         pabc::Vec_Int_t* vLits; // Dynamic vector of literals
 
-        static constexpr int NR_SIM_TTS = 32;
+        static const int NR_SIM_TTS = 32;
         std::vector<kitty::dynamic_truth_table> sim_tts { NR_SIM_TTS };
 
         bool fix_output_sim_vars(const spec& spec, int  t)
         {
+            if (spec.is_dont_care(0, t + 1)) {
+                return true;
+            }
             auto ilast_step = spec.nr_steps - 1;
             auto outbit = kitty::get_bit(
                 spec[spec.synth_func(0)], t + 1);
@@ -57,14 +60,14 @@ namespace percy
     public:
         const int OP_VARS_PER_STEP = 3;
 
-        knuth_fence2_encoder(solver_wrapper& solver)
+        ssv_fence2_encoder(solver_wrapper& solver)
         {
             // TODO: compute better upper bound on number of literals
             vLits = pabc::Vec_IntAlloc(128);
             set_solver(solver);
         }
 
-        ~knuth_fence2_encoder()
+        ~ssv_fence2_encoder()
         {
             pabc::Vec_IntFree(vLits);
         }
@@ -353,19 +356,19 @@ namespace percy
         }
 
         /// Ensure that each gate has 2 operands.
-        void create_fanin_clauses(const spec& spec)
+        bool create_fanin_clauses(const spec& spec)
         {
+            bool res = true;
             for (int i = 0; i < spec.nr_steps; i++) {
                 const auto nr_svars_for_i = nr_svars_for_step(spec, i);
                 for (int j = 0; j < nr_svars_for_i; j++) {
                     const auto sel_var = get_sel_var(spec, i, j);
                     pabc::Vec_IntSetEntry(vLits, j, pabc::Abc_Var2Lit(sel_var, 0));
                 }
-
-                const auto res = solver->add_clause(pabc::Vec_IntArray(vLits), 
+                res &= solver->add_clause(pabc::Vec_IntArray(vLits), 
                         pabc::Vec_IntArray(vLits) + nr_svars_for_i);
-                assert(res);
             }
+            return res;
         }
 
         /// Add clauses which ensure that every step is used at least once.
@@ -735,7 +738,9 @@ namespace percy
                 return false;
             }
 
-            create_fanin_clauses(spec);
+            if (!create_fanin_clauses(spec)) {
+                return false;
+            }
 
             if (spec.add_nontriv_clauses) {
                 create_nontriv_clauses(spec);
@@ -775,7 +780,9 @@ namespace percy
                 vcreate_tt_clauses(spec, t);
             }
             
-            create_fanin_clauses(spec);
+            if (!create_fanin_clauses(spec)) {
+                return false;
+            }
             create_cardinality_constraints(spec);
 
             if (spec.add_nontriv_clauses) {
@@ -879,7 +886,7 @@ namespace percy
             return sim_tts[spec.nr_in + spec.nr_steps - 1];
         }
 
-        void reset_sim_tts(int nr_in)
+        void reset_sim_tts(int nr_in) override 
         {
             for (int i = 0; i < NR_SIM_TTS; i++) {
                 sim_tts[i] = kitty::dynamic_truth_table(nr_in);